fix(auto-block): close the gaps /mr-review found in the shared-IP guards
Five findings, four of them ways a guard could quietly stop guarding.

- An invalid per-rule `mode` inherited the global mode instead of falling
  back to warn, so `'mode' => 'warm'` under an armed global silently
  blocked — contradicting the docblock directly above it.
- `shared_ip_user_threshold` was cast with `(int)`, and 0 is the "guard
  off" sentinel, so a blank or misspelled env value disabled the
  shared-IP guard silently. It now falls back to the default and logs;
  an explicit 0 still means off.
- The never_auto_block check compared the source enum strictly, but the
  model's cast accepts 'auto' as a string, so a caller passing one would
  have walked past the guard.
- The distinct-user query ran before the already-blocked filter, so it
  was computed every tick for addresses whose outcome was already known.
  Filtered first; still re-checked in the loop, because blocking one
  IPv6 address covers its prefix and can block a later offender mid-loop.

And two overclaims corrected rather than coded around:

- never_auto_block does NOT survive sync. A block another node's
  automation decided arrives as Sync, and the payload doesn't say where
  it came from, so the receiving node can't apply its own list. The
  CHANGELOG said "every automated path gets it", which was false. Filed
  as #56; README and CHANGELOG now say so plainly.
- The shared-IP guard counts signed-in users and cannot tell real ones
  from accounts an attacker registered. Documented as a limit, with
  never_block named as the control that isn't negotiable.

+++ config/watchtower.php
@@ -220,7 +220,10 @@ return [
     'auto_block' => [
         'enabled'                  => env('WATCHTOWER_AUTO_BLOCK_ENABLED', env('GUARD_AUTO_BLOCK_ENABLED', false)),
         'mode'                     => env('WATCHTOWER_AUTO_BLOCK_MODE', 'warn'),
-        'shared_ip_user_threshold' => (int) env('WATCHTOWER_SHARED_IP_USER_THRESHOLD', 3),
+        // Not cast here: 0 means "guard off", and a bare (int) would turn a
+        // blank or misspelled value into 0 too. The service parses it and
+        // falls back to the default, loudly, when it isn't a whole number.
+        'shared_ip_user_threshold' => env('WATCHTOWER_SHARED_IP_USER_THRESHOLD', 3),
         'block_duration_minutes'   => env('WATCHTOWER_AUTO_BLOCK_DURATION', env('GUARD_AUTO_BLOCK_DURATION', 60)),
         'rules'                    => [
             // Example — a rule you're still tuning. With no 'mode' it runs
